requirements.txt
Flask==2.1.2
requests==2.26.0
matplotlib==3.4.3
python-dotenv
pandas==1.3.3